Stop delegating targetAttachmentClass argument
- Avoids a deprecation error around overwriting computer properties
