+++ addon/templates/components/modal-dialog.hbs
@@ -21,7 +21,6 @@
   tetherClassPrefix=this.tetherClassPrefix
   constraints=this.constraints
   attachmentClass=this.attachmentClass
-  targetAttachmentClass=this.targetAttachmentClass
   stack=this.stack
   value=this.value
   onClickOverlay=this.onClickOverlayAction
